Revert "uploading results and code use in paper"
This reverts commit 2ca237d2598db9e07690e26647d4047ab2c7615f

+++ dtnsim/src/node/dtn/SdrModel.cc
@@ -24,7 +24,7 @@ void SdrModel::setNodesNumber(int nodesNumber)
 	this->nodesNumber_ = nodesNumber;
 }
 
-void SdrModel::setSize(double size)
+void SdrModel::setSize(int size)
 {
 	this->size_ = size;
 }
